Convert Pantone PMS codes to Munsell notation

Convert Pantone PMS codes to Munsell notation

PMS2Munsell(codes)

Arguments

codes
vector of PMS codes (e.g. '7630-c'), may contain NA, see pms.munsell.lut

Value

data.frame containing closest associated Munsell color (via rgb2munsell), hex notation, and perceptual color distance (dE00) between sRGB values and closest Munsell "chip".

Details

Conversion of Pantone spot colors (PMS code) is performed by look-up from pms.munsell.lut. Association is based on the "closest" Munsell color via CIE2000 distance metric (dE00) (see rgb2munsell). This is an experimental association between the two color systems and should not be used for precision color matching or paint mixing applications.

Possible uses include rough estimation of soil colors in the field, by means of color swatches based on the Pantone system. This type of color matching is most appropriate in an educational setting where official soil color books may be too expensive.

Note

Inspired by the work and outreach efforts of Dr. Karen Vaughan (UWY).

Author

D.E. Beaudette

Examples

# safely handles NA
codes <- c(NA, "7630-c", "102-c")

PMS2Munsell(codes)

#>     code     hex munsell      dE00
#> 1   <NA>    <NA>    <NA>        NA
#> 2 7630-c #633231  5R 3/4 2.9763935
#> 3  102-c #fce300 5Y 9/13 0.5202735