Ranking Systems for USDA Taxonomic Particle-Size and Substitute Classes of Mineral Soils

Generate a vector of USDA Particle-Size and Substitute Classes names, sorted according to approximate particle size

PSCS_levels()

Value

an ordered factor

References

Field Book for Describing and Sampling Soils, version 3.0

See also

Author

Stephen Roecker

Examples

# class codes
PSCS_levels()
#>  [1] fragmental                                              
#>  [2] pumiceous                                               
#>  [3] pumiceous or ashy-pumiceous over loamy-skeletal         
#>  [4] pumiceous or ashy-pumiceous over medial-skeletal        
#>  [5] cindery                                                 
#>  [6] pumiceous or ashy-pumiceous over medial                 
#>  [7] pumiceous or ashy-pumiceous over loamy                  
#>  [8] cindery over medial-skeletal                            
#>  [9] sandy-skeletal                                          
#> [10] cindery over medial                                     
#> [11] cindery over loamy                                      
#> [12] loamy-skeletal                                          
#> [13] loamy-skeletal over fragmental                          
#> [14] loamy-skeletal over cindery                             
#> [15] loamy-skeletal over gypseous-skeletal                   
#> [16] sandy-skeletal over loamy                               
#> [17] gypseous-skeletal                                       
#> [18] loamy-skeletal over sandy or sandy-skeletal             
#> [19] ashy-skeletal                                           
#> [20] ashy-skeletal over fragmental or cindery                
#> [21] loamy-skeletal over clayey                              
#> [22] ashy-skeletal over loamy-skeletal                       
#> [23] medial-skeletal                                         
#> [24] medial-skeletal over fragmental or cindery              
#> [25] ashy-skeletal over sandy or sandy-skeletal              
#> [26] medial-skeletal over loamy-skeletal                     
#> [27] ashy-pumiceous                                          
#> [28] medial-skeletal over sandy or sandy-skeletal            
#> [29] ashy-skeletal over clayey                               
#> [30] medial-pumiceous                                        
#> [31] clay-skeletal                                           
#> [32] sandy                                                   
#> [33] clayey-skeletal over sandy or sandy-skeletal            
#> [34] ashy                                                    
#> [35] ashy over clayey-skeletal                               
#> [36] ashy over pumiceous or cindery                          
#> [37] ashy over loamy-skeletal                                
#> [38] ashy over medial-skeletal                               
#> [39] sandy over loamy                                        
#> [40] coarse-loamy                                            
#> [41] coarse-loamy over fragmental                            
#> [42] ashy over sandy or sandy-skeletal                       
#> [43] sandy over clayey                                       
#> [44] ashy over medial                                        
#> [45] ashy over loamy                                         
#> [46] coarse-silty                                            
#> [47] coarse-loamy over sandy or sandy-skeletal               
#> [48] ashy over clayey                                        
#> [49] coarse-gypseous                                         
#> [50] coarse-silty over sandy or sandy-skeletal               
#> [51] coarse-loamy over clayey                                
#> [52] medial                                                  
#> [53] medial over clayey-skeletal                             
#> [54] medial over fragmental                                  
#> [55] medial over pumiceous or cindery                        
#> [56] coarse-silty over clayey                                
#> [57] medial over loamy-skeletal                              
#> [58] medial over ashy-pumiceous or ashy-skeletal             
#> [59] loamy                                                   
#> [60] loamy over pumiceous or cindery                         
#> [61] medial over sandy or sandy-skeletal                     
#> [62] medial over ashy                                        
#> [63] medial over loamy                                       
#> [64] fine-gypseous                                           
#> [65] medial over hydrous                                     
#> [66] loamy over sandy or sandy-skeletal                      
#> [67] loamy over ashy or ashy-pumiceous                       
#> [68] medial over clayey                                      
#> [69] loamy over coarse-gypseous                              
#> [70] loamy over fine-gypseous                                
#> [71] fine-loamy                                              
#> [72] fine-loamy over fragmental                              
#> [73] fine-silty                                              
#> [74] fine-silty over fragmental                              
#> [75] fine-loamy over sandy or sandy-skeletal                 
#> [76] hydrous                                                 
#> [77] hydrous over clayey-skeletal                            
#> [78] hydrous over fragmental                                 
#> [79] fine-silty over sandy or sandy-skeletal                 
#> [80] fine-loamy over clayey                                  
#> [81] hydrous over loamy-skeletal                             
#> [82] fine                                                    
#> [83] hydrous over sandy or sandy-skeletal                    
#> [84] fine-silty over clayey                                  
#> [85] hydrous over loamy                                      
#> [86] clayey                                                  
#> [87] clayey over fragmental                                  
#> [88] hydrous over clayey                                     
#> [89] clayey over loamy-skeletal                              
#> [90] clayey over gypseous-skeletal                           
#> [91] very-fine                                               
#> [92] clayey over sandy or sandy-skeletal                     
#> [93] clayey over coarse-gypseous                             
#> [94] clayey over loamy                                       
#> [95] clayey over fine-gypseous                               
#> [96] diatomaceous                                            
#> [97] pumiceous or ashy-pumiceous over sandy or sandy-skeletal
#> 97 Levels: fragmental < ... < pumiceous or ashy-pumiceous over sandy or sandy-skeletal