SoilProfileCollection Class

Basic class for storing soil profile collections, associated site data, and metadata.

Objects from the Class

Objects can be created by calls of the form new("SoilProfileCollection", ...).

Slots

idcol:

Object of class "character" the name of the column used to uniquely identify profiles

hzidcol:

Object of class "character" the name of the column used to uniquely identify horizons

depthcols:

Object of class "character" with the names of columns containing the horizon top and bottom boundaries

metadata:

Object of class "data.frame" with collection-level metadata, having a single row, and user-defined columns

horizons:

Object of class "data.frame" with 1 or more rows per profile

site:

Object of class "data.frame" with 1 row per profile

sp:

Object of class "SpatialPoints" with 1 row per profile

diagnostic:

Object of class "data.frame" with 0 or more rows per profile

restrictions:

Object of class "data.frame" with 0 or more rows per profile

Methods

$

signature(x = "SoilProfileCollection"): ...

$<-

signature(x = "SoilProfileCollection"): ...

[

signature(x = "SoilProfileCollection", i = "ANY", j = "ANY"): ...

[[

signature(x = "SoilProfileCollection", i = "ANY", j = "ANY"): ...

[[<-

signature(x = "SoilProfileCollection", i = "ANY", j = "ANY"): ...

coordinates<-

signature(object = "SoilProfileCollection"): ...

horizonDepths

signature(object = "SoilProfileCollection"): ...

horizons

signature(object = "SoilProfileCollection"): ...

horizons<-

signature(object = "SoilProfileCollection"): ...

idname

signature(object = "SoilProfileCollection"): ...

names

signature(x = "SoilProfileCollection"): ...

horizonNames

signature(object = "SoilProfileCollection"): ...

siteNames

signature(object = "SoilProfileCollection"): ...

length

signature(x = "SoilProfileCollection"): ...

max

signature(x = "SoilProfileCollection"): ...

metadata

signature(object = "SoilProfileCollection"): ...

metadata<-

signature(object = "SoilProfileCollection"): ...

min

signature(x = "SoilProfileCollection"): ...

profile_id

signature(object = "SoilProfileCollection"): ...

profile_plot

signature(object = "SoilProfileCollection"): ...

show

signature(object = "SoilProfileCollection"): ...

site

signature(object = "SoilProfileCollection"): ...

site<-

signature(object = "SoilProfileCollection"): ...

slab

signature(data = "SoilProfileCollection"): ...

units

signature(object = "SoilProfileCollection"): ...

units<-

signature(object = "SoilProfileCollection"): ...

Examples

# concatenate SoilProfileCollection objects
if (FALSE) {
require(plyr)
d <- ldply(1:10, random_profile)

# promote to SoilProfileCollection and plot
depths(d) <- id ~ top + bottom
plot(d)

# split into new SoilProfileCollection objects by index
d.1 <- d[1, ]
d.2 <- d[2, ]
d.345 <- d[3:5, ]

# recombine, note that profiles are sorted according to ID
d.new <- union(list(d.345, d.1, d.2))
plot(d.new)
}