Buntley-Westin (1965) Index

Calculate "Color Development Equivalent" by the method of Buntley & Westin (1965) "A Comparative Study of Developmental Color in a Chestnut-Chernozem-Brunizem Soil Climosequence" DOI: 10.2136/sssaj1965.03615995002900050029x. Originally developed for Mollisols, the Buntley-Westin index has been used as a tool to separate soils based on depth to particular colors.

buntley.westin.index(hue, chroma)

Arguments

hue
A character vector containing Munsell hues (e.g. "7.5YR")

chroma
A numeric vector containing Munsell chromas

Value

A numeric vector reflecting horizon redness (higher values = redder).