Dissolving horizon boundaries by grouping variables

This function dissolves or combines horizons share have a common set of grouping variables. It only combines those horizon records that are sequential (e.g. share a horizon boundary). Thus, it can be used to identify discontinuities in the grouping variables along a profile.

dissolve_hz(
  object,
  by,
  collapse = FALSE,
  id = "peiid",
  top = "hzdept",
  bot = "hzdepb",
  order = FALSE
)

Arguments

object

a data.frame

by

character: column names, to be used as grouping variables, within the object.

collapse

logical: indicating whether to not combine grouping variables before dissolving.

id

character: column name of the pedon ID within the object.

top

character: column name of the horizon top depth within the object.

bot

character: column name of the horizon bottom depth in the object.

order

logical: indicating whether or not to order the object by the id, top, and bot columns. #'

Value

A data.frame with the original id, by grouping variables, and non-consecutive horizon depths.

Details

This function assumes the profiles and horizons within the object follow the logic defined by checkHzDepthLogic (e.g. records are ordered sequentially by id, top, and bot and without gaps). If the records are not order, set the argument order = TRUE.

See also

Author

Stephen Roecker

Examples

# example data
data(jacobs2000)

spc <- jacobs2000

spc$dep_5 <- spc$depletion_pct >=5
spc$genhz <- generalize.hz(spc$name, c("A", "E", "B", "C"), c("A", "E", "B", "C")) 
h <- horizons(spc)

test <- dissolve_hz(h, by = c("genhz", "dep_5"), id = "id", top = "top", bot = "bottom")
#> non-character grouping variables are being converted to characters

head(h[c("id", "top", "bottom", "genhz", "dep_5")])
#>     id top bottom genhz dep_5
#> 1 92-1   0     18     A FALSE
#> 2 92-1  18     43     E FALSE
#> 3 92-1  43     79     B FALSE
#> 4 92-1  79    130     B FALSE
#> 5 92-1 130    153     B FALSE
#> 6 92-1 153    156     C FALSE
head(test)
#>     id top bottom   var val
#> 1 92-1   0     18 genhz   A
#> 2 92-1  18     43 genhz   E
#> 3 92-1  43    153 genhz   B
#> 4 92-1 153    213 genhz   C
#> 5 92-2   0     18 genhz   A
#> 6 92-2  18     46 genhz   E