Permute soil horizon depths using boundary distinctness

This method is most "believable" when used to _gently_ permute the data, on the order of moving boundaries a few centimeters in either direction. The nice thing about it is it can leverage semi-quantitative (ordered factor) levels of boundary distinctness/topography for the upper and lower boundary of individual horizons, given a set of assumptions to convert classes to a "standard deviation" (see example).

If you imagine a normal curve with its mean centered on the vertical (depth axis) at a RV horizon depth. By the Empirical Rule for Normal distribution, two "standard deviations" above or below that RV depth represent 95

So, a standard deviation of 1-2cm would yield a "boundary thickness" in the 3-5cm range ("clear" distinctness class).

Of course, boundaries are not symmetrical and this is at best an approximation for properties like organic matter, nutrients or salts that can have strong depth-dependence within horizons. Also, boundary topography is non-uniform. There are definitely ways to implement other distributions, but invokes more detailed assumptions about field data that are generally only semi-quantitative or are not available.

Future implementations may use boundary topography as a second hierarchical level (e.g. trig-based random functions), but think that distinctness captures the "uncertainty" about horizon separation at a specific "point" on the ground (or line in the profile quite well, and the extra variation may be hard to interpret, in general.

permute_profile(
  p,
  n = 100,
  id = NULL,
  boundary.attr,
  min.thickness = 1,
  soildepth = NULL,
  new.idname = "pID"
)

Arguments

p
A single-profile SoilProfileCollection

n
Number of permutations to generate (default: 100)

id
Over-rides n: a vector of (unique) profile IDs equal in length to number of permutations (n) to generate.

boundary.attr
Horizon attribute containing numeric "standard deviations" reflecting boundary transition distinctness

min.thickness
Minimum thickness of permuted horizons (default: 1)

soildepth
Depth below which horizon depths are not permuted (default: NULL)

new.idname
New column name to contain unique profile ID (default: pID)

Value

A SoilProfileCollection with n permutations of p.

Author

Andrew G. Brown

Examples

# # example with sp1 (using boundary distinctness)
data("sp1")
depths(sp1) <- id ~ top + bottom

# specify "standard deviation" for boundary thickness
#   consider a normal curve centered at boundary RV depth
# lookup table: ~maximum thickness of boundary distinctness classes, divided by 3
bound.lut <- c('V'=0.5,'A'=2,'C'=5,'G'=15,'D'=45) / 3

## V          A          C          G          D
## 0.1666667  0.6666667  1.6666667  5.0000000 15.0000000

sp1$bound_sd <- bound.lut[sp1$bound_distinct]

# hold any NA boundary distinctness constant
sp1$bound_sd[is.na(sp1$bound_sd)] <- 0

quantile(sp1$bound_sd, na.rm = TRUE)

#>        0%       25%       50%       75%      100% 
#> 0.0000000 0.0000000 0.6666667 0.6666667 1.6666667 

p <- sp1[3]