Plot soil profiles below a dendrogram

Plot soil profiles below a dendrogram, using methods from the ape package.

plotProfileDendrogram(
  x,
  clust,
  rotateToProfileID = FALSE,
  scaling.factor = 0.01,
  width = 0.1,
  y.offset = 0.1,
  dend.y.scale = max(clust$height * 2, na.rm = TRUE),
  dend.color = par("fg"),
  dend.width = 1,
  dend.type = c("phylogram", "cladogram"),
  debug = FALSE,
  ...
)

Arguments

x

a SoilProfileCollection object

clust

a hierarchical clustering object generated by hclust, cluster::agnes, or cluster::diana

rotateToProfileID

logical, attempt rotation of dendrogram according to original profile IDs, requires dendextend package

scaling.factor

vertical scaling of the profile heights (may have to tinker with this)

width

scaling of profile widths

y.offset

vertical offset for top of profiles

dend.y.scale

extent of y-axis (may have to tinker with this)

dend.color

dendrogram line color

dend.width

dendrogram line width

dend.type

dendrogram type, passed to plot.phylo(), either "phylogram" or "cladogram"

debug

logical, optionally print debugging data and return a data.frame of linking structure

...

additional arguments to plotSPC

Value

This function is typically called to create graphical output, when debug=TRUE a data.frame of IDs and linking structure used to build the figure.

Note

You may have to tinker with some of the arguments to get optimal arrangement and scaling of soil profiles.

Author

D.E. Beaudette