plot_distance_graph {aqp}	R Documentation

Between Individual Distance Plot

Description

Plot pair-wisee distances between individuals as line segments.

Usage

plot_distance_graph(D, idx = 1:dim(as.matrix((D)))[1])

Arguments

D
distance matrix, should be of class 'dist' or compatible class

idx
an integer sequence defining which individuals should be compared

Details

By default all individuals are plotting on the same axis. When there are more than about 10 individuals, the plot can become quite messy. See examples below for ideas.

Value

No value is returned.

Author(s)

Dylan E Beaudette

References

http://casoilresource.lawr.ucdavis.edu/

See Also

sp2, profile_compare

Examples

data(sp2)

d <- profile_compare(sp2, vars=c('prop','field_ph','hue','value'), 
max_d=100, k=0.01, sample_interval=5)

par(mfcol=c(3,1), mar=c(2.5,4.5,1,1))
plot_distance_graph(d, idx=1:6)
plot_distance_graph(d, idx=7:12)
plot_distance_graph(d, idx=12:18)

[Package aqp version 1.9.3 Index]