Repair Problematic Lower Horizon Depths

Attempt a simple repair of horizon bottom depths in the presence of NA, or in cases where the horizon shares a common top and bottom depth. Both situations are common in pedon description where "contact" (Cd, Cr, R, etc.) was described without a lower depth. This repair is only applied to the deepest horizon within a profile as identified by getLastHorizonID.

repairMissingHzDepths(x, adj = 10)

Arguments

x

SoilProfileCollection

adj

vertical offset applied to "repair" affected bottom depths

Value

SoilProfileCollection with a new (logical) horizon-level attribute .repaired marking affected horizons