Textural conversions

These functions consist of several conversions between sand, silt and clay to texture class and visa versa, textural modifers to rock fragments, and grain size composition to the family particle size class.

texcl_to_ssc(texcl = NULL, clay = NULL)

ssc_to_texcl(sand = NULL, clay = NULL, as.is = FALSE, droplevels = TRUE)

texmod_to_fragvoltot(texmod, lieutex = NULL)

texture_to_taxpartsize(texcl = NULL, clay = NULL, sand = NULL, fragvoltot = NULL)

Arguments

sand
vector of sand percentages (default: NULL)

clay
vector of clay percentages (default: NULL)

fragvoltot
vector of rock fragment percentages (default: NULL)

texcl
vector of texture classes than conform to the USDA code conventions (e.g. c|C, sil|SIL, sl|SL, cos|COS)

texmod
vector of textural modifers that conform to the USDA code convenctions (e.g. gr|GR, grv|GRV)

lieutex
vector of in lieu of texture terms that conform to the USDA code convenctions (e.g. gr|GR, pg|PG), only used when fragments or artifacts are > 90 percent by volume (default: NULL))

as.is
logical: should character vectors be converted to factors? (default: TRUE)

droplevels
logical: indicating whether to drop unused levels in factors. This is useful when the results have a large number of unused classes, which can waste space in tables and figures.

Details

These functions are intended to impute missing values or check existing values. The ssc_to_texcl() function uses the same logic as the particle size estimator calculation in NASIS to classify sand and clay into texture class. The results are stored in data(soiltexture) and used by texcl_to_ssc() as a lookup table to convert texture class to sand, silt and clay. The function texcl_to_ssc() replicates the functionality described by Levi (2017).

Unlike the other functions, texture_to_taxpartsize() is intended to be computed on weighted averages within the family particle size control section. Also recall from the criteria that carbonate clay should be subtracted from clay content and added to silt content. Similarily, if the percent of very fine sand is known it should be substracted from the sand, and added to the silt content.

References

Matthew R. Levi, Modified Centroid for Estimating Sand, Silt, and Clay from Soil Texture Class, Soil Science Society of America Journal, 2017, 81(3):578-588, ISSN 1435-0661, https://doi.org/10.2136/sssaj2016.09.0301.

Value

A vector with the results.

Examples

# \donttest{

# example of ssc_to_texcl()
tex <- expand.grid(sand = 0:100, clay = 0:100)
tex <- subset(tex, (sand + clay) < 101)
tex$texcl <- ssc_to_texcl(sand = tex$sand, clay = tex$clay)
head(tex)

#>   sand clay texcl
#> 1    0    0    si
#> 2    1    0    si
#> 3    2    0    si
#> 4    3    0    si
#> 5    4    0    si
#> 6    5    0    si

# example of texcl_to_ssc(texcl)
texcl <- c("cos", "s", "fs", "vfs", "lcos", "ls",
"lfs", "lvfs", "cosl", "sl", "fsl", "vfsl", "l",
"sil", "si", "scl", "cl", "sicl", "sc", "sic", "c"
)
test <- texcl_to_ssc(texcl)
head(test <- cbind(texcl, test), 10)

#>    texcl sand silt clay
#> 1    cos   92    5    3
#> 2      s   92    5    3
#> 3     fs   92    5    3
#> 4    vfs   92    5    3
#> 5   lcos   82   12    6
#> 6     ls   82   12    6
#> 7    lfs   82   12    6
#> 8   lvfs   82   12    6
#> 9   cosl   65   25   10
#> 10    sl   65   25   10

# example of texcl_to_ssc(texcl, clay)
data(soiltexture)
st <- soiltexture$values
idx <- sample(1:length(st$texcl), 10)
st <- st[idx, ]
ssc <- texcl_to_ssc(texcl = st$texcl)
head(cbind(texcl = st$texcl, clay = ssc$clay))

#>      texcl clay
#> [1,] "scl" "27"
#> [2,] "sil" "13"
#> [3,] "sil" "13"
#> [4,] "c"   "55"
#> [5,] "sl"  "10"
#> [6,] "c"   "55"

# example of texmod_to_fragvoltol()
frags <- c("gr", "grv", "grx", "pgr", "pgrv", "pgrx")
head(texmod_to_fragvoltot(frags))

#>   texmod fragvoltot_l fragvoltot_r fragvoltot_h fragvoltot_h_nopf
#> 1     gr           15           25           34                34
#> 2    grv           35           48           59                59
#> 3    grx           60           75           89                89
#> 4    pgr           15           25           34                 0
#> 5   pgrv           35           48           59                 0
#> 6   pgrx           60           75           89                 0
#>   fragvoltot_r_nopf fragvoltot_l_nopf
#> 1                25                15
#> 2                48                35
#> 3                75                60
#> 4                 0                 0
#> 5                 0                 0
#> 6                 0                 0

# example of texture_to_taxpartsize()
tex <- data.frame(texcl = c("c", "cl", "l", "ls", "s"),
                  clay  = c(55, 33, 18, 6, 3),
                  sand  = c(20, 33, 42, 82, 93),
                  fragvoltot = c(35, 15, 34, 60, 91))
tex$fpsc<-texture_to_taxpartsize(texcl = tex$texcl, clay = tex$clay,sand = tex$sand, fragvoltot = tex$fragvoltot)
head(tex)

#>   texcl clay sand fragvoltot            fpsc
#> 1     c   55   20         35 clayey-skeletal
#> 2    cl   33   33         15      fine-loamy
#> 3     l   18   42         34      fine-loamy
#> 4    ls    6   82         60  loamy-skeletal
#> 5     s    3   93         91      fragmental

# example of texture_to_taxpartsize() with carbonate clay and very fine sand
carbclay <- rnorm(5, 2, 3)
vfs <- rnorm(5, 10, 3)
st$fpsc <- texture_to_taxpartsize(texcl = tex$texcl, clay = tex$clay - carbclay, sand = tex$sand - vfs, fragvoltot = tex$fragvoltot)

#> Warning: some records are > 100% or < 0%, or the calcuated silt fraction is > 100% or < 0%

#> Warning: some of the texcl records don't match the calculated texcl via ssc_to_texcl()

head(tex)

#>   texcl clay sand fragvoltot            fpsc
#> 1     c   55   20         35 clayey-skeletal
#> 2    cl   33   33         15      fine-loamy
#> 3     l   18   42         34      fine-loamy
#> 4    ls    6   82         60  loamy-skeletal
#> 5     s    3   93         91      fragmental

# }