unroll {aqp}	R Documentation

Unroll Genetic Horizons

Description

Generate a descretized vector of genetic horizons along a user-defined pattern.

Usage

unroll(top, bottom, prop, max_depth, bottom_padding_value = NA, strict=FALSE)

Arguments

top
vector of upper horizon boundaries, must be an integer

bottom
vector of lower horizon boundaries, must be an integer

prop
vector of some property to be "unrolled" over a regular sequence

max_depth
maximum depth to which missing data is padded with NA

bottom_padding_value
value to use when padding missing data

strict
should horizons be strictly checked for self-consistency? defaults to FALSE

Details

This function is used internally by several higher-level components of the aqp package. Basic error checking is performed to make sure that bottom and top horizon boundaries make sense. Note that the horizons should be sorted according to depth before using this function. The max_depth argument is used to specifiy the maximum depth of profiles within a collection, so that data from any profile shallower than this depth is padded with NA.

Value

a vector of "unrolled" property values

Author(s)

Dylan E. Beaudette

References

http://casoilresource.lawr.ucdavis.edu/

Examples


data(sp1)

# subset a single soil profile:
sp1.1 <- subset(sp1, subset=id == 'P001')

## Error: comparison (1) is possible only for atomic and list types


# demonstrate how this function works
x <- with(sp1.1, unroll(top, bottom, prop, max_depth=50))

## Error: object 'sp1.1' not found

plot(x, 1:length(x), ylim=c(90,0), type='b', cex=0.5)   

## Error: formal argument "type" matched by multiple actual arguments



[Package aqp version 1.4-7 Index]