SoilProfileCollection-class {aqp}	R Documentation

SoilProfileCollection Class

Description

Basic class for storing soil profile collections, associated site data, and metadata.

Objects from the Class

Objects can be created by calls of the form new("SoilProfileCollection", ...).

Slots

idcol:

Object of class "character" the name of the column used to uniquely identify profiles

depthcols:

Object of class "character" with the names of columns containing the horizon top and bottom boundaries

metadata:

Object of class "data.frame" with collection-level metadata, having a single row, and user-defined columns

horizons:

Object of class "data.frame" with 1 or more rows per profile

site:

Object of class "data.frame" with 1 row per profile

sp:

Object of class "SpatialPoints" with 1 row per profile

diagnostic:

Object of class "data.frame" with 0 or more rows per profile

Methods

$

signature(x = "SoilProfileCollection"): ...

$<-

signature(x = "SoilProfileCollection"): ...

[

signature(x = "SoilProfileCollection", i = "ANY", j = "missing"): ...

[[

signature(x = "SoilProfileCollection", i = "ANY", j = "missing"): ...

[[<-

signature(x = "SoilProfileCollection", i = "ANY", j = "missing"): ...

coordinates<-

signature(object = "SoilProfileCollection"): ...

horizonDepths

signature(object = "SoilProfileCollection"): ...

horizons

signature(object = "SoilProfileCollection"): ...

horizons<-

signature(object = "SoilProfileCollection"): ...

idname

signature(object = "SoilProfileCollection"): ...

names

signature(x = "SoilProfileCollection"): ...

horizonNames

signature(object = "SoilProfileCollection"): ...

siteNames

signature(object = "SoilProfileCollection"): ...

length

signature(x = "SoilProfileCollection"): ...

max

signature(x = "SoilProfileCollection"): ...

metadata

signature(object = "SoilProfileCollection"): ...

metadata<-

signature(object = "SoilProfileCollection"): ...

min

signature(x = "SoilProfileCollection"): ...

profile_id

signature(object = "SoilProfileCollection"): ...

profile_plot

signature(object = "SoilProfileCollection"): ...

show

signature(object = "SoilProfileCollection"): ...

site

signature(object = "SoilProfileCollection"): ...

site<-

signature(object = "SoilProfileCollection"): ...

slab

signature(data = "SoilProfileCollection"): ...

units

signature(object = "SoilProfileCollection"): ...

units<-

signature(object = "SoilProfileCollection"): ...

Author(s)

Pierre Roudier and Dylan E. Beaudette

Examples

# concatenate SoilProfileCollection objects
## Not run: 
##D require(plyr)
##D d <- ldply(1:10, random_profile)
##D 
##D # promote to SoilProfileCollection and plot
##D depths(d) <- id ~ top + bottom
##D plot(d)
##D 
##D # split into new SoilProfileCollection objects by index
##D d.1 <- d[1, ]
##D d.2 <- d[2, ]
##D d.345 <- d[3:5, ]
##D 
##D # recombine, note that profiles are sorted according to ID
##D d.new <- rbind(d.345, d.1, d.2)
##D plot(d.new)
## End(Not run)

## Not run: 
##D # these next examples should throw an error
##D # insert a missing horizon boundary
##D data(sp1)
##D sp1$top[1] <- NA
##D depths(sp1) <- id ~ top + bottom
## End(Not run)

[Package aqp version 1.9.1 Index]