Generate a vector of USDA soil texture codes or class names, sorted according to approximate particle size

Generate a vector of USDA soil texture codes or class names, sorted according to approximate particle size

SoilTextureLevels(which = "codes")

Arguments

which
'codes' (texture codes) or 'names' (texture class names)

Value

an ordered factor

References

Field Book for Describing and Sampling Soils, version 3.0

Examples

# class codes
SoilTextureLevels()

#>  [1] cos  s    fs   vfs  lcos ls   lfs  lvfs cosl sl   fsl  vfsl l    sil  si  
#> [16] scl  cl   sicl sc   sic  c   
#> 21 Levels: cos < s < fs < vfs < lcos < ls < lfs < lvfs < cosl < sl < ... < c

# class names
SoilTextureLevels(which = 'names')

#>  [1] coarse sand          sand                 fine sand           
#>  [4] very fine sand       loamy coarse sand    loamy sand          
#>  [7] loamy fine sandy     loamy very fine sand coarse sandy loam   
#> [10] sandy loam           fine sandy loam      very fine sandy loam
#> [13] loam                 silt loam            silt                
#> [16] sandy clay loam      clay loam            silty clay loam     
#> [19] sandy clay           silty clay           clay                
#> 21 Levels: coarse sand < sand < fine sand < ... < clay