aggregateSoilDepth {aqp}	R Documentation

Probabalistic Estimation of Soil Depth

Description

Estimate the most-likely depth to contact within a collection of soil profiles.

Usage


aggregateSoilDepth(x, groups, crit.prob = 0.9, name = "hzname", p = "Cr|R|Cd", ...)

Arguments

x
a SoilProfileCollection object

groups
the name of a site-level attribute that defines groups of profiles within a collection

crit.prob
probability cuttoff used to determine where the most likely depth to contact will be, e.g. 0.9 translates to 90% of profiles are shallower than this depth

name
horizon-level attribute where horizon designation is stored

p
a REGEX pattern that matches non-soil genetic horizons

...
additional arguments to slab

Details

This function computes a probability-based estimate of soil depth by group. If no grouping variable exists, a dummy value can be used to compute a single estimate. The crit.prob argument sets the critical probability (e.g. 0.9) at which soil depth within a group of profiles is determined. For example, a crit.prob of 0.95 might result in an estimated soil depth (e.g. 120cm) where 95% of the profiles (by group) had depths that were less than or equal to 120cm.

Value

A data.frame is returned, with as many rows as there are unique group labels, as specified in groups.

Author(s)

D.E. Beaudette

See Also

estimateSoilDepth,slab

Examples

data(sp1)
depths(sp1) <- id ~ top + bottom
site(sp1) <- ~ group

aggregateSoilDepth(sp1, 'group', crit.prob = 0.9)

##   group soil.top soil.bottom
## 1     1        0         232
## 2     2        0          67

[Package aqp version 1.9.1 Index]