aqp-package {aqp}	R Documentation

Algorithms for Quantitative Pedology

Description

The aqp (Algorithms for Quantitative Pedology) package for R was developed to address some of the difficulties associated with processing soils information, specifically related to visualization, aggregation, and classification of soil profile data. This package is based on a mix of S3/S4 functions and classes, and most functions use basic dataframes as input, where rows represent soil horizons and columns define properties of those horizons. Common to most functions are the requirements that horizon boundaries are defined as depth from 0, and that profiles are uniquely defined by an id column. The aqp package defines an S4 class, "SoilProfileCollection", for storage of profile-level metadata, as well as summary, print, and plotting methods that have been customized for common tasks related to soils data.

Demos: demo(aqp)

demo(slope_effect_hz_thickness)

Extended Examples: AQP-related blog posts

Introduction to SoilProfileCollection objects

Soil profile dissimilarity demo

Vertical vs. perpendicular horizon depth measurement demo

Author(s)

Dylan E. Beaudette <debeaudette@ucdavis.edu>

References

http://casoilresource.lawr.ucdavis.edu/

See Also

ca630, sp1, sp2, sp3, sp4, sp5

[Package aqp version 1.6-1 Index]