Check a SoilProfileCollection object for errors in horizon depths.

This function inspects a SoilProfileCollection object, looking for 4 common errors in horizon depths: 1) bottom depths shallower than top depths, 2) equal top and bottom depths, 3) missing top or bottom depths (e.g. NA), and, 4) gaps or overlap between adjacent horizons.

checkHzDepthLogic(x)

Arguments

x
a SoilProfileCollection object

Details

This function replaces test_hz_logic, now marked as deprecated.

Value

A data.frame with as many rows as profiles in x.

id

Profile IDs, named according to idname(x)

depthLogic

boolean, errors related to depth logic

sameDepth

boolean, errors related to same top/bottom depths

missingDepth

boolean, NA in top / bottom depths

overlapOrGap

boolean, gaps or overlap in adjacent horizons

valid

boolean, profile passes all tests

Note

There is currently no simple way to fix errors identified by this function. Stay tuned for a fixHzDepthErrors().

Examples

## sample data
data(sp3)
depths(sp3) <- id ~ top + bottom

# these data should be clean
(res <- checkHzDepthLogic(sp3))

#>    id depthLogic sameDepth missingDepth overlapOrGap valid
#> 1   1      FALSE     FALSE        FALSE        FALSE  TRUE
#> 2  10      FALSE     FALSE        FALSE        FALSE  TRUE
#> 3   2      FALSE     FALSE        FALSE        FALSE  TRUE
#> 4   3      FALSE     FALSE        FALSE        FALSE  TRUE
#> 5   4      FALSE     FALSE        FALSE        FALSE  TRUE
#> 6   5      FALSE     FALSE        FALSE        FALSE  TRUE
#> 7   6      FALSE     FALSE        FALSE        FALSE  TRUE
#> 8   7      FALSE     FALSE        FALSE        FALSE  TRUE
#> 9   8      FALSE     FALSE        FALSE        FALSE  TRUE
#> 10  9      FALSE     FALSE        FALSE        FALSE  TRUE