estimateSoilDepth {aqp}	R Documentation

Estimate Soil Depth

Description

Estimate the soil depth of a single profile within a SoilProfileCollection object.

Usage

estimateSoilDepth(f, name = "hzname", top = "hzdept", bottom = "hzdepb", p = 'Cr|R|Cd', 
no.contact.depth = NULL, no.contact.assigned = NULL)

Arguments

f
A SoilProfileCollection object of length 1, e.g. a single profile

name
the name of the column that contains horizon designations

top
the name of the column that contains horizon top depths

bottom
the name of the column that contains horizon bottom depths

p
a REGEX pattern for determining contact with bedrock

no.contact.depth
in the absense of contact with bedrock, a depth at which we can assume a standard depth

no.contact.assigned
assumed standard depth

Value

a single integer, the soil depth

Author(s)

D.E. Beaudette and J.M. Skovlin

See Also

getSoilDepthClass, profileApply

Examples

data(sp1)
depths(sp1) <- id ~ top + bottom

# apply to each profile in a collection, and save as site-level attribute
sp1$depth <- profileApply(sp1, estimateSoilDepth, name='name', top='top', bottom='bottom')

## Not run: 
# sample data
data(gopheridge, package='soilDB')

# run on a single profile
estimateSoilDepth(gopheridge[1, ])

# apply to an entire collection
profileApply(gopheridge, estimateSoilDepth)

## End(Not run)

[Package aqp version 1.9.3 Index]