generalize.hz {aqp}	R Documentation

Generalize Horizon Names

Description

Generalize a vector of horizon names, based on new classes, and REGEX patterns.

Usage

generalize.hz(x, new, pat, non.matching.code)

Arguments

x
a character vector of horizon names

new
a character vector of new horizon classes

pat
a character vector of REGEX, same length as x

non.matching.code
text used to describe any horizon not matching any item in pat

Value

factor of the same length as x

Author(s)

Dylan E. Beaudette

References

http://casoilresource.lawr.ucdavis.edu/

Examples

data(sp1)

# check original distribution of hz designations
table(sp1$name)

# generalize
sp1$genhz <- generalize.hz(sp1$name, 
new=c('O','A','B','C','R'), 
pat=c('O', '^A','^B','C','R'))

# see how we did / what we missed
table(sp1$genhz, sp1$name)

[Package aqp version 1.9.3 Index]