get.ml.hz {aqp}	R Documentation

Determine ML Horizon Boundaries

Description

This function accepts input from slab() along with a vector of horizon names, and returns a data.frame of the most likely horizonat boundaries.

Usage


get.ml.hz(x, hz.names)

Arguments

x
output from slab

hz.names
vector of horizon names

Details

This function is expecting that x is a data.frame containing columns identified in hz.names.

Value

A dataframe with the following columns:

hz
horizon names

top
top boundary

bottom
bottom boundary

confidence
integrated probability / ML horizon thickness, rounded to the nearest integer

Author(s)

D.E. Beaudette

See Also

slab

Examples


data(sp1)
depths(sp1) <- id ~ top + bottom

# normalize horizon names: result is a factor
sp1$name <- generalize.hz(sp1$name, 
  new=c('O','A','B','C'), 
  pat=c('O', '^A','^B','C'))

# compute slice-wise probability so that it sums to contributing fraction, from 0-150
a <- slab(sp1, fm= ~ name, cpm=1, slab.structure=0:150)

# generate table of ML horizonation
get.ml.hz(a, c('O','A','B','C'))

##   hz top bottom confidence
## 4  O   0      1         28
## 1  A   1     31         75
## 2  B  31    144         57
## 3  C 144    149         67



[Package aqp version 1.5-1 Index]