getSoilDepthClass {aqp}	R Documentation

Generate Soil Depth Class Matrix

Description

Generate a boolean matrix of soil depth classes from a SoilProfileCollection object.

Usage

getSoilDepthClass(f, depth.classes = c(very.shallow = 25, 
shallow = 50, mod.deep = 100, deep = 150, very.deep = 1000), ...)

Arguments

f
a SoilProfileCollection object

depth.classes
a named vector of classes and depth breaks

...
arguments passed to estimateSoilDepth

Value

a data.frame containing soil depth and depth class for each profile, see examples

Author(s)

D.E. Beaudette and J.M. Skovlin

See Also

estimateSoilDepth

Examples

data(sp1)
depths(sp1) <- id ~ top + bottom

# generate depth-class matrix
sdc <- getSoilDepthClass(sp1, name='name', top='top', bottom='bottom')

# inspect
head(sdc)

# join back into sp1 as site-level data
site(sp1) <- sdc

## Not run: 
# sample data
data(gopheridge, package='soilDB')

getSoilDepthClass(gopheridge)

## End(Not run)

[Package aqp version 1.9.10 Index]