hzDistinctnessCodeToOffset {aqp}	R Documentation

Convert Horizon Distinctness Codes

Description

This function accepts a vector of horizon distinctness codes, a look-up vector of codes, and a corresponding vector of vertical offsets. The defaults are based on USDA-NCSS field methods.

Usage

hzDistinctnessCodeToOffset(x, codes = c("A", "C", "G", "D"), offset = c(0.5, 1.5, 5, 10))

Arguments

x
vector of distinctness codes to be converted to offsets

codes
vector of unique distinctness codes

offset
vector of offsets associated with distinctness codes

Details

Missing data (NA) or codes that are not defined in the ‘codes’ argument are returned as 0 offsets.

Value

a vector of vertical offsets, with the same length as the number of distinctness codes passed to the function

Author(s)

D.E. Beaudette

References

http://www.nrcs.usda.gov/wps/portal/nrcs/detail/soils/ref/?cid=nrcs142p2_054184

See Also

plotSPC

Examples

data(sp1)
hzDistinctnessCodeToOffset(sp1$bound_distinct)

[Package aqp version 1.9.10 Index]