Mix Munsell Colors via Spectral Library

Simulate subtractive mixing of colors in Munsell notation, similar to the way in which mixtures of pigments operate.

mixMunsell(x, w = rep(1, times = length(x))/length(x), n = 1)

Arguments

x
vector of colors in Munsell notation

w
vector of proportions, can sum to any number

n
number of closest matching color chips

Value

a data.frame with the closest matching Munsell color(s)

Details

Pending

Note

This functions is slower than soilDB::estimateColorMixture() (weighted average of colors in CIELAB coordinates) but more accurate.

References

inspiration / calculations based on the work of Scott Burns: https://arxiv.org/ftp/arxiv/papers/1710/1710.06364.pdf

related discussion on Stack Overflow: https://stackoverflow.com/questions/10254022/implementing-kubelka-munk-like-krita-to-mix-colours-color-like-paint/29967630#29967630

spectral library source: https://www.munsellcolourscienceforpainters.com/MunsellResources/SpectralReflectancesOf2007MunsellBookOfColorGlossy.txt

See also

Author

D.E. Beaudette