Plot soil profiles below a dendrogram or cladogram

Plot soil profiles below a dendrogram, using methods from the ape package.

Usage

plotProfileDendrogram(
  x,
  clust,
  rotateToProfileID = FALSE,
  scaling.factor = 5 * mean(clust$height)/max(x),
  width = 0.25,
  y.offset = 60 * mean(clust$height)/max(x),
  dend.y.scale = max(clust$height * 2, na.rm = TRUE),
  dend.color = par("fg"),
  dend.width = 1,
  dend.type = c("phylogram", "cladogram"),
  debug = FALSE,
  ...
)

Arguments

x

a SoilProfileCollection object

clust

a hierarchical clustering object generated by hclust(), cluster::agnes(), or cluster::diana()

rotateToProfileID

logical, attempt rotation of dendrogram according to original profile IDs, requires dendextend package

scaling.factor

numeric, vertical scaling of the profile heights (default estimate printed if not specified)

width

numeric, scaling of profile widths

y.offset

numeric, vertical offset for top of profiles (default estimate printed if not specified)

dend.y.scale

numeric, scaling of dendrogram (default estimate printed if not specified)

dend.color

dendrogram line color

dend.width

dendrogram line width

dend.type

dendrogram type, passed to plot.phylo(), either "phylogram" or "cladogram"

debug

logical, optionally print debugging data and return a data.frame of linking structure

...

additional arguments to plotSPC()

Value

This function is typically called to create graphical output, when debug = TRUE a data.frame of IDs and linking structure used to build the figure.

Note

The default values of scaling.factor, y.offset, and dend.y.scale are based on clust$height and max(x), and should be close to optimal. Use the defaults as a starting point for finer control.

Author

D.E. Beaudette