Quickly assess relative state of site and horizon slots

Determine "state" of SoilProfileCollection before or after major modifications of site or horizon slot contents.

Three logical checks are performed on the site and horizon tables, and a fourth attribute valid returns TRUE when all three checks are TRUE.

Check 1: Site order is reflected in horizons. Which means: mo intermingling of sites within horizon table. Each profile is a block of contiguous rows in the horizon table. Though, they _may_ be out of order within profiles.

Check 2: Site IDs match coalesced profile ID from horizon. This check ensures the same _relative_ ordering, but horizons still may be out of order within profiles

Check 3: Horizon IDs are in order of profile ID in site and within profiles, have correct top-depth sequence.

spc_in_sync(object)

Arguments

object
A SoilProfileCollection

Value

data.table